updates based on feedback; renames view icons; adds in Model and Metrics icons
