--- src/lib/components/icons/AlertOctagon.svelte
@@ -1,18 +1,0 @@
-<script>
-  export let size = "1em";
-  export let color = "currentColor";
-</script>
-
-<svg
-  height={size}
-  viewBox="0 0 24 24"
-  fill="none"
-  xmlns="http://www.w3.org/2000/svg"
->
-  <path
-    fill-rule="evenodd"
-    clip-rule="evenodd"
-    d="M8.32843 3C7.79797 3 7.28931 3.21069 6.91418 3.58582L3.58582 6.91418C3.21069 7.28931 3 7.79797 3 8.32843V15.6716C3 16.202 3.21069 16.7107 3.58582 17.0858L6.91418 20.4142C7.28931 20.7893 7.79797 21 8.32843 21H15.6716C16.202 21 16.7107 20.7893 17.0858 20.4142L20.4142 17.0858C20.7893 16.7107 21 16.202 21 15.6716V8.32843C21 7.79797 20.7893 7.28931 20.4142 6.91418L17.0858 3.58582C16.7107 3.21069 16.202 3 15.6716 3H8.32843ZM11.0029 12.2501C11.0029 12.3882 11.1148 12.5 11.2529 12.5H13.088C13.2261 12.5 13.3381 12.388 13.338 12.2499L13.3353 7.24988C13.3352 7.11182 13.2233 7 13.0853 7H11.2501C11.112 7 11.0001 7.112 11.0001 7.25012L11.0029 12.2501ZM11.0001 16.25C11.0001 16.3881 11.1121 16.5 11.2501 16.5H13.0853C13.2234 16.5 13.3353 16.3881 13.3353 16.25V14.4203C13.3353 14.2822 13.2234 14.1703 13.0853 14.1703H11.2501C11.1121 14.1703 11.0001 14.2822 11.0001 14.4203V16.25Z"
-    fill={color}
-  />
-</svg>
