--- src/lib/components/icons/Code.svelte
@@ -1,16 +1,0 @@
-<script>
-  export let size = "1em";
-  export let color = "currentColor";
-</script>
-
-<svg
-  height={size}
-  viewBox="0 0 24 24"
-  fill="none"
-  xmlns="http://www.w3.org/2000/svg"
->
-  <path
-    d="M7.37522 16.7815L8.62523 15.2194L4.60122 12.0005L8.62523 8.78148L7.37522 7.21947L2.37522 11.2195C2.25807 11.3132 2.1635 11.432 2.0985 11.5672C2.03351 11.7024 1.99976 11.8505 1.99976 12.0005C1.99976 12.1505 2.03351 12.2986 2.0985 12.4338C2.1635 12.569 2.25807 12.6878 2.37522 12.7815L7.37522 16.7815ZM16.6252 7.21947L15.3752 8.78148L19.3993 12.0005L15.3752 15.2194L16.6252 16.7815L21.6253 12.7815C21.7424 12.6878 21.8369 12.569 21.9019 12.4338C21.967 12.2986 22.0007 12.1505 22.0007 12.0005C22.0007 11.8505 21.967 11.7024 21.9019 11.5672C21.8369 11.432 21.7424 11.3132 21.6253 11.2195L16.6252 7.21947ZM14.9762 3.21648L10.9762 21.2164L9.02322 20.7825L13.0232 2.78247L14.9762 3.21648Z"
-    fill={color}
-  />
-</svg>
