+++ src/lib/components/icons/Explore.svelte
@@ -1,0 +1,22 @@
+<script>
+  export let size = "1em";
+  export let color = "currentColor";
+</script>
+
+<svg
+  height={size}
+  viewBox="0 0 24 24"
+  fill="none"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    d="M4 4.5C4 4.22386 3.77614 4 3.5 4H2.5C2.22386 4 2 4.22386 2 4.5V19.5C2 19.7761 2.22386 20 2.5 20H21.5C21.7761 20 22 19.7761 22 19.5V18.5C22 18.2239 21.7761 18 21.5 18H4V4.5Z"
+    fill={color}
+  />
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M21.5205 5.89513C21.7158 6.0904 21.7158 6.40698 21.5205 6.60224L14.5205 13.6022C14.3252 13.7975 14.0087 13.7975 13.8134 13.6022L11.1447 10.9335L8.16767 13.9105C7.97241 14.1058 7.65582 14.1058 7.46056 13.9105L6.75346 13.2034C6.55819 13.0081 6.55819 12.6915 6.75346 12.4963L10.7911 8.45863C10.9864 8.26336 11.303 8.26336 11.4982 8.45863L14.167 11.1274L20.1063 5.18803C20.3016 4.99276 20.6181 4.99276 20.8134 5.18803L21.5205 5.89513Z"
+    fill={color}
+  />
+</svg>
