+++ src/lib/components/icons/Metrics.svelte
@@ -1,0 +1,24 @@
+<script>
+  export let size = "1em";
+  export let color = "currentColor";
+</script>
+
+<svg
+  height={size}
+  viewBox="0 0 24 24"
+  fill="none"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M12.4335 5.14772C12.152 5.00716 11.8208 5.00724 11.5395 5.14793L5.36016 8.23758C5.02137 8.40697 4.80737 8.75324 4.80737 9.13202L4.80746 14.8408C4.80746 15.2196 5.02147 15.5659 5.36026 15.7352L11.5452 18.8276C11.8267 18.9683 12.1581 18.9683 12.4396 18.8276L18.6282 15.7333C18.967 15.5639 19.181 15.2176 19.181 14.8388L19.1809 9.13586C19.1809 8.75693 18.9667 8.41056 18.6277 8.24125L12.4335 5.14772ZM10.645 3.35908C11.4892 2.937 12.4827 2.93677 13.3271 3.35846L19.5213 6.45198C20.5383 6.9599 21.1809 7.99903 21.1809 9.13582L21.181 14.8388C21.181 15.9751 20.539 17.0139 19.5227 17.5221L13.334 20.6164C12.4895 21.0387 11.4954 21.0387 10.6508 20.6164L4.46587 17.5241C3.4495 17.016 2.80747 15.9772 2.80746 14.8409L2.80737 9.13205C2.80736 7.99572 3.44936 6.9569 4.46573 6.44872L10.645 3.35908Z"
+    fill={color}
+  />
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M7.24885 9.11693C7.37203 8.86979 7.67224 8.76929 7.91938 8.89247L11.9932 10.9228L16.0561 8.89518C16.3032 8.77187 16.6035 8.87221 16.7268 9.1193L17.1263 9.91988C17.2496 10.167 17.1493 10.4672 16.9022 10.5905L12.9408 12.5675V16.5262C12.9408 16.8024 12.717 17.0262 12.4408 17.0262H11.5461C11.2699 17.0262 11.0461 16.8024 11.0461 16.5262V12.5678L7.07421 10.5883C6.82706 10.4651 6.72656 10.1649 6.84974 9.91772L7.24885 9.11693Z"
+    fill="black"
+  />
+</svg>
