--- src/lib/components/icons/MetricsConfig.svelte
@@ -1,31 +1,0 @@
-<script>
-  export let size = "1em";
-  export let color = "currentColor";
-</script>
-
-<svg
-  height={size}
-  viewBox="0 0 24 24"
-  fill="none"
-  xmlns="http://www.w3.org/2000/svg"
->
-  <path
-    fill-rule="evenodd"
-    clip-rule="evenodd"
-    d="M12.2285 18.8704C12.0987 18.6493 12.1572 18.3658 12.364 18.2142L13.1654 17.6263C13.0958 17.2152 13.0939 16.7917 13.1646 16.373L12.364 15.7859C12.1573 15.6342 12.0987 15.3507 12.2285 15.1296L12.9739 13.8596C13.101 13.6431 13.3689 13.5544 13.6001 13.6523L14.5303 14.0462C14.8635 13.7767 15.2373 13.5665 15.6341 13.4202L15.7513 12.4406C15.7814 12.1892 15.9946 12 16.2478 12L17.7522 12C18.0054 12 18.2186 12.1892 18.2487 12.4406L18.3658 13.4196C18.5624 13.4925 18.7556 13.582 18.9435 13.6887C19.1314 13.7955 19.3068 13.9153 19.4692 14.0463L20.3999 13.6522C20.6311 13.5544 20.899 13.6431 21.0261 13.8596L21.7715 15.1296C21.9013 15.3507 21.8428 15.6342 21.636 15.7858L20.8346 16.3736C20.9043 16.7848 20.9061 17.2084 20.8355 17.627L21.636 18.2142C21.8428 18.3659 21.9013 18.6493 21.7715 18.8705L21.0261 20.1405C20.899 20.357 20.6311 20.4457 20.3999 20.3478L19.4697 19.9539C19.1365 20.2234 18.7627 20.4336 18.3658 20.5798L18.2487 21.5594C18.2186 21.8108 18.0054 22 17.7522 22L16.2478 22C15.9946 22 15.7814 21.8108 15.7513 21.5594L15.6342 20.5803C15.4376 20.5075 15.2444 20.418 15.0565 20.3113C14.8686 20.2046 14.6932 20.0847 14.5308 19.9537L13.6001 20.3478C13.3689 20.4456 13.101 20.3569 12.9739 20.1404L12.2285 18.8704ZM15.9536 18.7831C16.9546 19.3516 18.2347 19.0142 18.8126 18.0295C19.3906 17.0448 19.0476 15.7856 18.0466 15.2171C17.0455 14.6486 15.7654 14.9859 15.1875 15.9707C14.6095 16.9554 14.9525 18.2145 15.9536 18.7831Z"
-    fill={color}
-  />
-  <rect x="7" y="8" width="10" height="2" rx="0.5" fill="black" />
-  <path
-    d="M7 12C7 11.7239 7.22386 11.5 7.5 11.5H10.5C10.7761 11.5 11 11.7239 11 12V13C11 13.2761 10.7761 13.5 10.5 13.5H7.5C7.22386 13.5 7 13.2761 7 13V12Z"
-    fill={color}
-  />
-  <path
-    d="M7 15.5C7 15.2239 7.22386 15 7.5 15H8.5C8.77614 15 9 15.2239 9 15.5V16.5C9 16.7761 8.77614 17 8.5 17H7.5C7.22386 17 7 16.7761 7 16.5V15.5Z"
-    fill={color}
-  />
-  <path
-    d="M3 4.5C3 4.22386 3.22386 4 3.5 4H20.5C20.7761 4 21 4.22386 21 4.5V9.5C21 9.77614 20.7761 10 20.5 10H19.5C19.2239 10 19 9.77614 19 9.5V6H5V19H8.5C8.77614 19 9 19.2239 9 19.5V20.5C9 20.7761 8.77614 21 8.5 21H3.5C3.22386 21 3 20.7761 3 20.5V4.5Z"
-    fill={color}
-  />
-</svg>
