+++ src/lib/components/icons/Model.svelte
@@ -1,0 +1,16 @@
+<script>
+  export let size = "1em";
+  export let color = "currentColor";
+</script>
+
+<svg
+  height={size}
+  viewBox="0 0 24 24"
+  fill="none"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    d="M6.98483 16.4692C7.20044 16.6417 7.51506 16.6067 7.68757 16.3912L8.31277 15.6099C8.48533 15.3942 8.45037 15.0795 8.23471 14.907L4.60122 12.0005L8.23471 9.09388C8.45037 8.92135 8.48532 8.60666 8.31276 8.39102L7.68758 7.6098C7.51506 7.39422 7.20045 7.35928 6.98484 7.53177L2.37522 11.2195C2.25807 11.3132 2.1635 11.432 2.0985 11.5672C2.03351 11.7024 1.99976 11.8505 1.99976 12.0005C1.99976 12.1505 2.03351 12.2986 2.0985 12.4338C2.1635 12.569 2.25807 12.6878 2.37522 12.7815L6.98483 16.4692ZM17.0156 7.53177C16.8 7.35929 16.4854 7.39422 16.3128 7.60981L15.6877 8.39102C15.5151 8.60666 15.5501 8.92135 15.7658 9.09387L19.3993 12.0005L15.7657 14.907C15.5501 15.0795 15.5151 15.3942 15.6877 15.6099L16.3128 16.3911C16.4853 16.6067 16.8 16.6417 17.0156 16.4692L21.6253 12.7815C21.7424 12.6878 21.8369 12.569 21.9019 12.4338C21.967 12.2986 22.0007 12.1505 22.0007 12.0005C22.0007 11.8505 21.967 11.7024 21.9019 11.5672C21.8369 11.432 21.7424 11.3132 21.6253 11.2195L17.0156 7.53177ZM14.4881 3.10801C14.7577 3.16792 14.9277 3.43501 14.8678 3.70457L11.0847 20.7283C11.0248 20.9979 10.7577 21.1679 10.4881 21.108L9.51133 20.8909C9.24176 20.831 9.07178 20.5639 9.13169 20.2944L12.9148 3.27057C12.9747 3.001 13.2418 2.83103 13.5113 2.89094L14.4881 3.10801Z"
+    fill={color}
+  />
+</svg>
