+++ src/routes/dev/icons.svelte
@@ -11,11 +11,12 @@
 
   const viewIcons = Object.keys(icons).filter(
     (i) =>
-      i.includes("Code") ||
-      i.includes("LineChart") ||
-      i.includes("Parquet") ||
-      i.includes("MetricsConfig")
+      i.includes("Source") ||
+      i.includes("Model") ||
+      i.includes("Metrics") ||
+      i.includes("Explore")
   );
+  viewIcons.reverse();
 
   const others = Object.keys(icons).filter(
     (i) =>
@@ -25,10 +26,15 @@
 
 <h1 class="text-xl mb-4">Iconography</h1>
 
-<div class="flex flex-row flex-wrap w-screen gap-x-16">
+<!-- <div class="flex flex-row flex-wrap w-screen gap-x-16"> -->
+<div
+  style="
+    column-width: 300px;
+  "
+>
   {#each [["view icons", viewIcons], ["form icons", formIcons], ["data types", typeIcons], ["others", others]] as [name, iconSet]}
-    <div>
-      <h2 class="mb-1 mt-3">{name}</h2>
+    <div style="break-inside: avoid;">
+      <h2 class="mb-1 mt-2">{name}</h2>
       <div
         class="grid gap-x-3 gap-y-2 items-center"
         style:grid-template-columns="max-content max-content max-content"
@@ -37,10 +43,10 @@
           <div style:width="200px">
             {iconName.split("/").slice(-1)[0]}
           </div>
-          <div style:font-size="2rem">
+          <div style:font-size="1.25rem">
             <svelte:component this={icons[iconName].default} />
           </div>
-          <div style:font-size="1rem">
+          <div style:font-size=".95rem">
             <svelte:component this={icons[iconName].default} />
           </div>
         {/each}
@@ -48,22 +54,3 @@
     </div>
   {/each}
 </div>
-<!-- <div
-  class="grid gap-x-3 gap-y-2 items-center"
-  style:grid-template-columns="max-content max-content max-content max-content"
->
-  {#each Object.keys(icons) as iconName}
-    <div>
-      {iconName.split("/").slice(-1)[0]}
-    </div>
-    <div style:font-size="2rem">
-      <svelte:component this={icons[iconName].default} />
-    </div>
-    <div style:font-size="1rem">
-      <svelte:component this={icons[iconName].default} />
-    </div>
-    <div style:font-size="12px">
-      <svelte:component this={icons[iconName].default} />
-    </div>
-  {/each}
-</div> -->
